requirements-test.txt
twine
pytest
requests
pytest-cov>=2.4.0,<2.6
coverage
codecov
pep8
autopep8
httmock
alchemy-mock
mock==2.0.0
testing.postgresql
fakeredis
